bits: negative skips are not implemented

+++ src/util/bits.c
@@ -182,7 +182,7 @@ uint32_t bs_read( BITSTREAM *bs, int i_count )
     return bb_read(&bs->bb, i_count);
 }
 
-void bb_skip( BITBUFFER *bb, ssize_t i_count )
+void bb_skip( BITBUFFER *bb, size_t i_count )
 {
     bb->i_left -= i_count;
 
@@ -194,10 +194,10 @@ void bb_skip( BITBUFFER *bb, ssize_t i_count )
     }
 }
 
-void bs_skip( BITSTREAM *bs, ssize_t i_count )
+void bs_skip( BITSTREAM *bs, size_t i_count )
 {
     int left;
-    int bytes = i_count >> 3;
+    size_t bytes = i_count >> 3;
 
     if (bs->bb.p + bytes >= bs->bb.p_end) {
         bs->pos = bs->pos + (bs->bb.p - bs->bb.p_start);
